requirements.txt
assemblyai
python-dotenv
moviepy
streamlit
anthropic
